fix: rename helper module for GitHub Pages
