+++ js/donationHelpers.js
